requirements.txt
jax[tpu]==0.7.2
flax==0.11.2
orbax-checkpoint==0.11.24
tqdm
torch
torchvision
hydra-core
wandb